refactor(consumer-rhi): consolidate VulkanLayout newtype into one canonical home
Promote `VulkanLayout` into `streamlib-consumer-rhi::vulkan_layout` —
the single shared crate every surface adapter already depends on —
and have `streamlib-adapter-vulkan` and `streamlib-adapter-cpu-readback`
re-export from there. Both adapters previously carried near-duplicate
`pub struct VulkanLayout(pub i32)` definitions with subset constants;
post-#572 cdylib tightening promoted cpu-readback's copy from
`pub(crate)` to `pub`, which made the duplication public-API surface.

The unified definition carries the union of constants both adapters
were using (UNDEFINED / GENERAL / COLOR_ATTACHMENT_OPTIMAL /
SHADER_READ_ONLY_OPTIMAL / TRANSFER_SRC_OPTIMAL / TRANSFER_DST_OPTIMAL).
Internal `vk()` accessor renamed to `as_vk()` to match Rust's
`as_<target>` cast-method convention now that it's public; the three
in-tree adapter callsites are updated.
