+++ libs/streamlib-adapter-cpu-readback/src/lib.rs
@@ -38,7 +38,8 @@ pub use adapter::{
     InProcessCpuReadbackCopyTrigger, TriggerPlane,
 };
 pub use context::CpuReadbackContext;
-pub use state::{HostSurfaceRegistration, VulkanLayout};
+pub use state::HostSurfaceRegistration;
+pub use streamlib_consumer_rhi::VulkanLayout;
 pub use view::{
     CpuReadbackPlaneView, CpuReadbackPlaneViewMut, CpuReadbackReadView, CpuReadbackWriteView,
 };
