+++ libs/streamlib-adapter-vulkan/src/lib.rs
@@ -24,5 +24,6 @@ mod view;
 pub use adapter::VulkanSurfaceAdapter;
 pub use context::VulkanContext;
 pub use raw_handles::{raw_handles, RawVulkanHandles};
-pub use state::{HostSurfaceRegistration, VulkanLayout};
+pub use state::HostSurfaceRegistration;
+pub use streamlib_consumer_rhi::VulkanLayout;
 pub use view::{VulkanReadView, VulkanWriteView};
